requirements.txt
langchain_ollama
langchain_core
langchain
langgraph
langchain_aws
botocore
python-dotenv
ipykernel
twelvedata
pandas
MetaTrader5
langchain_openai
fastapi
uvicorn
pymongo
matplotlib
scipy